ref: split init/sync and slice import_shared (#165)
Three monolithic CLI files mixed unrelated concerns and made review
harder than it needed to be. Splitting by concern keeps each file
focused on a single job and matches the file-level convention already
established in adapter packages.

- init.go now holds only cobra wiring; scaffold and preset logic move
  to init_scaffold.go and init_preset.go.
- sync.go now holds only cobra wiring; orchestration, JSON output,
  and gitignore-flag helpers move to sync_run.go. Filter logic was
  already extracted in filter.go.
- import_shared.go splits into import_shared_md.go (markdown slicing
  and frontmatter), import_shared_mcp.go (JSON->YAML MCP helpers),
  and import_shared_tree.go (recursive copy with header strip). The
  base file keeps the next-steps printer and the YAML coercion shim.
  Helpers stay in the cli package because Go's internal-package rule
  blocks cli from importing internal/adapters/internal/emit; importer
  code is not adapter-shared logic.
