+++ internal/cli/import_shared_tree.go
@@ -1,0 +1,58 @@
+package cli
+
+import (
+	"fmt"
+	"io/fs"
+	"os"
+	"path/filepath"
+
+	"github.com/chemaclass/agnostic-ai/internal/adapters/header"
+)
+
+// copyDirTree walks srcDir recursively and writes every regular file
+// byte-for-byte into the matching location under dstDir, recreating
+// the directory layout as it goes. File mode bits are preserved so an
+// executable script remains executable on the destination. Symlinks
+// are not followed; if they appear inside a skill folder they are
+// silently skipped (skills are documented to be plain files +
+// directories — symlinks would not survive a tar/zip release anyway).
+func copyDirTree(srcDir, dstDir string) error {
+	return filepath.WalkDir(srcDir, func(path string, d fs.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+		rel, err := filepath.Rel(srcDir, path)
+		if err != nil {
+			return err
+		}
+		target := filepath.Join(dstDir, rel)
+		if d.IsDir() {
+			return os.MkdirAll(target, 0o755)
+		}
+		if !d.Type().IsRegular() {
+			return nil
+		}
+		info, err := d.Info()
+		if err != nil {
+			return err
+		}
+		data, err := os.ReadFile(path)
+		if err != nil {
+			return fmt.Errorf("read %s: %w", path, err)
+		}
+		// Strip the agnostic-ai provenance header from SKILL.md so a
+		// roundtrip (claude / codex emit -> import) does not bake the
+		// header into the source spec. Sibling assets pass through
+		// byte-for-byte because they are user-authored.
+		if filepath.Base(path) == "SKILL.md" {
+			data = []byte(header.Strip(string(data)))
+		}
+		if err := os.MkdirAll(filepath.Dir(target), 0o755); err != nil {
+			return fmt.Errorf("mkdir %s: %w", filepath.Dir(target), err)
+		}
+		if err := os.WriteFile(target, data, info.Mode().Perm()); err != nil {
+			return fmt.Errorf("write %s: %w", target, err)
+		}
+		return nil
+	})
+}
